+++ src/half_edge.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, Clone)]
+struct Mesh {
+    vertex_edge: Vec<(usize, usize)>,
+    edge_face: Vec<(usize, usize)>,
+}
